Fixed some gravity-issues and attempted to implement jumping
Gravity is only Not added, if player is on the ground.  So if the player is less gravity.y above the ground/platform, then  the pos will recieve some gravity, and player will be a little submerged in the ground/platform. The higher gravity had become, the farther into the ground the player would go.

Also started on implementing the jumping-mechanic, but it's not done yet.