Wanted to optimise collision - Processing didn't
Wanted to make a new function to get all the different color-values that player needs for collision with walls/blocks (and call said function as the first things every frame). This would make if-statements a little shorter, and I wanted to implement some more pixels to check color for, to refine the collision. The function broke jumping when I tried to add it to just the 4-line if-statement in Player class (line 81-84), and I can only think of Processing being weird again... :/