I have no clue what the problem is
I'm trying to print the different things in the console, but I'm not getting any smarter from it. I'm also starting to question if I even know how to code anymore, or if it has all just been more or less blind luck that I ever got to some code that worked (on the surface, if nothing else)?