Tried fixing the new method for verticleMovement() - didn't work but got interesting results
Fix didn't work. Copied the entire function and switched out all instances of "m.colorAt()" with a corresponding "pixel_" instead, just for good measure.
Narrowing down problem and found that for-loop seems to be the problem (maybe). Apparently it never enters the if-statement with the break, nor does the loop's condition ever become false; 'pos.y' increases mad fast and apparently the player's feet are always in the air (somehow??).