IT FINALLY WORKS - and added some small comments
Finally got verticleMovement() to work - for-loop kept going, because the color-values (the "pixel_" variables) never got updated in the loop, so it looked at the colors at the player's feet  right as player was in the air, and the "if (pixelHalfFrame ...)" triggered. Now color-values are updated throughout the function (which looks a little ugly), so falling doesn't break the program.