Changed some things about variables in Player class
Changed some variables out for others, changed types and more of that sort of stuff. Mostly due to this being more logical and sort of paves a better way of letting player fly out of portals with the speed he had when entering a portal.
I say "logical" because only the x-value of the old 'vel' PVector was used (specifically for sideways movement with 'a' and 'd'). All such vectors are now just floats.
Also 'initialGrravity' is just (0, 0) instead of (0, 0.2) because 'vel' should just be nothing, when there's no need to add it to player's pos.

Added a bunch of comments.
I am working on making an IntList with the color of black and yellow, so if-statements can become shorter, but I'm not done with it yet, and need to push *before* I break something...