Worked on better collision - which we (sort of) have now, along with some teleportation stuff
The game should have decent enough collision now (regarding player running into walls and falling onto platforms). There is, however, a problem if the player jumps onto a platform (seems to happen if player is moving sideways and lands). Then the foot-level pixel on the side which the player was going sinks into the platform, meaning the player has to jump up to continue moving in the desired direction. It has only been observed when only 1 foot lands on the platform, and the player can therefore still move the opposite direction just fine.

There has also been a few small changes to some function names, since parts of the original were moved to a new function - done so to avoid unnecessary code being run in a function that is called often in the program.

There was added a function for detecting player's collision with portals, since this would be necessary in order for player to be able to shoot out of portals.