FLYING OUT OF PORTALS WORKS! (also fixed some minor things here and there)
The player can now fly out of portals. Accomplished by saving the direction a portal is facing, then when the player is falling and is about to hit the ground in the next frame, then the code checks if player is going to hit a portal. If so, redirects player's velocity according to which way the portal, that the player teleports to, is facing, so player flies out of the portal.

Also commented some more things, along with fixing a few existing comments.

That player can fly out of portals did cause an issue on stage 2, where player would fly through the barrier.  Temporarily fixed by making the barrier slightly less black and adding this new color to bullet's collision code. Bullet can therefore still hit the barrier (though player can walk right through)