Tweaked button hitbox by the top exit and had the stage changed a little
Stage 2 has been changed back to almost original. The barrrier is now black again (instead of near-black) and the platform the player aims at is now positioned so player doesn't fly through barrier, when jumping into a portal to continue on.
Removed the extra collision for bullet, since the "near-black" check is obsolete now.